api/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
gunicorn==20.1.0
idna==2.10
manage.py==0.2.10
mysqlclient==2.0.3
Pillow==8.1.0
PyJWT==1.7.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.52.0
urllib3==1.26.3
whitenoise==5.2.0
